+++ src/decoder.rs
@@ -1,161 +1,9 @@
-use std::vec::Vec;
-use std::cell::RefCell;
 use std::rc::Rc;
-use std::cmp;
-use rand::{Rng, sample, StdRng, SeedableRng};
-
-use soliton::IdealSoliton;
-
-#[derive(Clone, Debug)]
-pub enum EncoderType {
-    /// The first k symbols of a systematic Encoder correspond to the first k source symbols
-    /// In case there is no loss, no repair needed. After the first k symbols are sent, it continous
-    /// like in the Random case.
-    Systematic,
-    /// Begins immediately with random encoding.
-    /// This may be a better choice when used with high-loss channels.
-    Random,
-}
-
-/// Encoder for Luby transform codes
-#[derive(Clone)]
-pub struct Encoder {
-    data: Vec<u8>,
-    len: usize,
-    blocksize: usize,
-    rng: StdRng,
-    cnt_blocks: usize,
-    sol: IdealSoliton,
-    pub cnt: usize,
-    encodertype: EncoderType,
-}
-
-#[derive(Debug)]
-pub enum DropType {
-    /// First is seed, second degree
-    Seeded(usize, usize),
-    /// Just a list of edges
-    Edges(usize),
-}
-
-/// A Droplet is created by the Encoder.
-#[derive(Debug)]
-pub struct Droplet {
-    /// The droptype can be based on seed or a list of edges
-    pub droptype: DropType,
-    /// The payload of the Droplet
-    pub data: Vec<u8>,
-}
-
-impl Droplet {
-    fn new(droptype: DropType, data: Vec<u8>) -> Droplet {
-        Droplet {
-            droptype: droptype,
-            data: data,
-        }
-    }
-}
-
-impl Encoder {
-    /// Constructs a new encoder for Luby transform codes.
-    /// In case you send the packages over UDP, the blocksize
-    /// should be the MTU size.
-    ///
-    /// There are two versions of the 'Encoder', Systematic and Random.
-    /// The Systematic encoder first produces a set of the source symbols. After each
-    /// symbol is sent once, it switches to Random.
-    ///
-    /// The Encoder implements the iterator. You can use the iterator
-    /// to produce an infinte stream of Droplets
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// extern crate rand;
-    /// extern crate fountaincode;
-    ///
-    /// fn main() {
-    ///     use fountaincode::ltcode::{Encoder, EncoderType};
-    ///     use self::rand::{thread_rng, Rng};
-    ///
-    ///     let s:String = thread_rng().gen_ascii_chars().take(1_024).collect();
-    ///     let buf = s.into_bytes();
-    ///
-    ///     let mut enc = Encoder::new(buf, 64, EncoderType::Random);
-    ///
-    ///     for i in 1..10 {
-    ///         println!("droplet {:?}: {:?}", i, enc.next());
-    ///     }
-    /// }
-    /// ```
-    pub fn new(data: Vec<u8>, blocksize: usize, encodertype: EncoderType) -> Encoder {
-        let mut rng = StdRng::new().unwrap();
-
-        let len = data.len();
-        let cnt_blocks = ((len as f32) / blocksize as f32).ceil() as usize;
-        let sol = IdealSoliton::new(cnt_blocks, rng.gen::<usize>());
-        Encoder {
-            data: data,
-            len: len,
-            blocksize: blocksize,
-            rng: rng,
-            cnt_blocks: cnt_blocks,
-            sol: sol,
-            cnt: 0,
-            encodertype: encodertype,
-        }
-    }
-}
-
-fn get_sample_from_rng_by_seed(seed: usize, n: usize, degree: usize) -> Vec<usize> {
-    let seedarr: &[_] = &[seed];
-    let mut rng: StdRng = SeedableRng::from_seed(seedarr);
-    sample(&mut rng, 0..n, degree)
-}
-
-impl Iterator for Encoder {
-    type Item = Droplet;
-    fn next(&mut self) -> Option<Droplet> {
-        let drop = match self.encodertype {
-            EncoderType::Random => {
-                let degree = self.sol.next().unwrap() as usize; //TODO: try! macro
-                let seed = self.rng.gen::<u32>() as usize;
-                let sample = get_sample_from_rng_by_seed(seed, self.cnt_blocks, degree);
-                let mut r: Vec<u8> = vec![0; self.blocksize];
-
-                for k in sample {
-                    let begin = k * self.blocksize;
-                    let end = cmp::min((k + 1) * self.blocksize, self.len);
-                    let mut j = 0;
-
-                    for i in begin..end {
-                        r[j] ^= self.data[i];
-                        j += 1;
-                    }
-                }
-                Some(Droplet::new(DropType::Seeded(seed, degree), r))
-            }
-            EncoderType::Systematic => {
-                let begin = self.cnt * self.blocksize;
-                let end = cmp::min((self.cnt + 1) * self.blocksize, self.len);
-                let mut r: Vec<u8> = vec![0; self.blocksize];
-
-                let mut j = 0;
-                for i in begin..end {
-                    r[j] = self.data[i];
-                    j += 1;
-                }
-                if (self.cnt + 2) > self.cnt_blocks {
-                    self.encodertype = EncoderType::Random;
-                }
-                Some(Droplet::new(DropType::Edges(self.cnt), r))
-            }
-        };
-
-        self.cnt += 1;
-        drop
-    }
-}
+use std::cell::RefCell;
+use crate::types::{DropType, CatchResult};
+use crate::block::Block;
+use crate::droplet::{Droplet, RxDroplet};
+use crate::encoder::get_sample_from_rng_by_seed;
 
 /// Decoder for the Luby transform
 pub struct Decoder {
@@ -176,26 +24,6 @@ pub struct Statistics {
     pub unknown_chunks: usize,
 }
 
-#[derive(Debug)]
-pub enum CatchResult {
-    Finished(Vec<u8>, Statistics),
-    Missing(Statistics),
-}
-
-#[derive(Debug)]
-struct RxDroplet {
-    edges_idx: Vec<usize>,
-    data: Vec<u8>,
-}
-
-struct Block {
-    idx: usize,
-    edges: Vec<Rc<RefCell<RxDroplet>>>,
-    begin_at: usize,
-    is_known: bool,
-}
-
-
 impl Decoder {
     /// Creates a new Decoder for LT codes
     ///
@@ -206,8 +34,9 @@ impl Decoder {
     /// extern crate fountaincode;
     ///
     /// fn main() {
-    ///     use fountaincode::ltcode::{Encoder, EncoderType, Decoder};
-    ///     use fountaincode::ltcode::CatchResult::*;
+    ///     use self::fountaincode::encoder::Encoder;
+    ///     use self::fountaincode::decoder::Decoder;
+    ///     use self::fountaincode::types::*;
     ///     use self::rand::{thread_rng, Rng};
     ///
     ///     let s:String = thread_rng().gen_ascii_chars().take(1_024).collect();
@@ -240,12 +69,7 @@ impl Decoder {
         let data: Vec<u8> = vec![0; number_of_chunks * blocksize];
         let mut edges: Vec<Block> = Vec::with_capacity(number_of_chunks);
         for i in 0..number_of_chunks {
-            let blk = Block {
-                idx: i,
-                edges: Vec::new(),
-                begin_at: blocksize * i,
-                is_known: false,
-            };
+            let blk = Block::new(i, Vec::new(), blocksize * i, false);
             edges.push(blk);
         }
