+++ src/lib.rs
@@ -1,4 +1,8 @@
 extern crate rand;
 
 pub mod soliton;
-pub mod ltcode;
+pub mod droplet;
+pub mod block;
+pub mod encoder;
+pub mod decoder;
+pub mod types;
